move toFixedNumber() to common functions; specify tandem temp basal rate in significant digits
