+++ lib/commonFunctions.js
@@ -23,6 +23,15 @@ var api = require('./core/api.js');
 var rollbar =require('../app/utils/rollbar');
 var isBrowser = typeof window !== 'undefined';
 
+Number.prototype.toFixedNumber = function(decimalPlaces, truncating){
+  var pow = Math.pow(10, decimalPlaces);
+  if(truncating) {
+     return +( Math.trunc(this*pow) / pow );
+  } else {
+    return +( Math.round(this*pow) / pow );
+  }
+};
+
 /**
  * Computes the number of milliseconds after midnight on the date specified.
  *
