+++ lib/drivers/animas/animasLogic.js
@@ -15,14 +15,7 @@
  * == BSD2 LICENSE ==
  */
 
- Number.prototype.toFixedNumber = function(decimalPlaces, truncating){
-   var pow = Math.pow(10, decimalPlaces);
-   if(truncating) {
-      return +( Math.trunc(this*pow) / pow );
-   } else {
-     return +( Math.round(this*pow) / pow );
-   }
- };
+ var common = require('../../commonFunctions');
 
  var roundAndNotNegative = function(number) {
    if(number < 0) {
