+++ lib/drivers/tandem/tandemDriver.js
@@ -666,10 +666,6 @@ module.exports = function (config) {
   // Tandem represents values with false precision (https://en.wikipedia.org/wiki/False_precision),
   // so we have to specify the number of significant digits
   var SIGNIFICANT_DIGITS = 5;
-  Number.prototype.toFixedNumber = function(significant){
-    var pow = Math.pow(10,significant);
-    return +( Math.round(this*pow) / pow );
-  };
 
   var pending = false;    //if serial send is pending, we have to wait before we send again
   var INTERVAL_FREQ = 5;  // time between sending serial packets
