+++ lib/drivers/tandem/tandemSimulator.js
@@ -142,7 +142,7 @@ exports.make = function(config){
           };
 
           tempBasal.suppressed = suppressed;
-          tempBasal.rate = suppressed.rate * tempBasal.percent;
+          tempBasal.rate = (suppressed.rate * tempBasal.percent).toFixedNumber(5);
           // actual temp basal duration is based on basal rate change event (temp_rate_end)
           tempBasal.duration = Date.parse(event.time) - Date.parse(tempBasal.time);
           delete tempBasal.time_left;
